Move post-condition unit tests from `stackslib` -> `clarity`
